% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_blank_filtering.R
\name{blank_calc_ratio}
\alias{blank_calc_ratio}
\title{Calculate the sample / blank ratio}
\usage{
blank_calc_ratio(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible).
}
\description{
Calculate sample / blank ratio..
}
