% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{cleanup}
\alias{cleanup}
\title{Cleanup import MSDIAL results}
\usage{
cleanup(data_df = NULL)
}
\arguments{
\item{data_df}{data.frame, containing the raw imported MSIDAL results.}
}
\value{
cleaned up data.frame
}
\description{
Cleanup import MSDIAL results.
}
