% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImport.R
\name{DataImport}
\alias{DataImport}
\title{R6 Data import class}
\description{
Data import class. This will be the parent class for several other classes.
}
\references{
Dunn et al. Nature Protocols 6, 1060-1083 (2011)
}
\author{
Rico Derks

Yassene Mohammed

E. Nevedomskaya
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character(1), containing a name.}

\item{\code{files}}{list(), containing all files.}

\item{\code{tables}}{list() containing all data tables.}

\item{\code{indices}}{list() containing all kind of indices.}

\item{\code{params}}{list(), containing all kind of parameter settings for the methods.}

\item{\code{history}}{data.frame() with the history of all steps.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataImport-new}{\code{DataImport$new()}}
\item \href{#method-DataImport-print}{\code{DataImport$print()}}
\item \href{#method-DataImport-show_data_info}{\code{DataImport$show_data_info()}}
\item \href{#method-DataImport-set_files}{\code{DataImport$set_files()}}
\item \href{#method-DataImport-import}{\code{DataImport$import()}}
\item \href{#method-DataImport-set_parameters}{\code{DataImport$set_parameters()}}
\item \href{#method-DataImport-calc_qcpool}{\code{DataImport$calc_qcpool()}}
\item \href{#method-DataImport-median_bc}{\code{DataImport$median_bc()}}
\item \href{#method-DataImport-loess_bc}{\code{DataImport$loess_bc()}}
\item \href{#method-DataImport-combat_bc}{\code{DataImport$combat_bc()}}
\item \href{#method-DataImport-blank_filtering}{\code{DataImport$blank_filtering()}}
\item \href{#method-DataImport-plot_qc_rsd}{\code{DataImport$plot_qc_rsd()}}
\item \href{#method-DataImport-plot_qc_trend}{\code{DataImport$plot_qc_trend()}}
\item \href{#method-DataImport-clone}{\code{DataImport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-new}{}}}
\subsection{Method \code{new()}}{
Initialization function for DataImport class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$new(name = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character(1), name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-print}{}}}
\subsection{Method \code{print()}}{
Print a nice summary of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used at the moment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-show_data_info"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-show_data_info}{}}}
\subsection{Method \code{show_data_info()}}{
Show information about the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$show_data_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-set_files"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-set_files}{}}}
\subsection{Method \code{set_files()}}{
Set the file names for data and meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$set_files(data_files = NULL, meta_file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_files}}{character(), containing the full file names.}

\item{\code{meta_file}}{character(1), containing the full meta file name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-import"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-import}{}}}
\subsection{Method \code{import()}}{
This method imports all data and meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-set_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-set_parameters}{}}}
\subsection{Method \code{set_parameters()}}{
Method to set all parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$set_parameters(
  ids = NULL,
  columns = NULL,
  regex = NULL,
  rsd = NULL,
  imputation = NULL,
  batch_correction = NULL,
  blank_filtering = NULL,
  normalization = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{list(), column names containing information about, sample id,
feature id. Valid entries are: id_col_meta, id_col_data.}

\item{\code{columns}}{list(), column names containing information about,
sample type, group column, batch column. Valid entries are: id_col_meta, 
id_col_sample, type_column, group_column, batch_column, order_column.}

\item{\code{regex}}{list(), regular expressions to recognize blanks, qcs, qcpools and samples.
Valid entries are blanks, qcs, qcpools and samples.}

\item{\code{rsd}}{list(), set rsd filtering parameters.}

\item{\code{imputation}}{list(), imputation parameters.}

\item{\code{batch_correction}}{list(), batch correction parameters.}

\item{\code{blank_filtering}}{list(), blank filtering parameters.}

\item{\code{normalization}}{list(), normalization parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Run this function first before importing the data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-calc_qcpool"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-calc_qcpool}{}}}
\subsection{Method \code{calc_qcpool()}}{
Run the QCpool calculations. Calculate RSD and Trend of all features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$calc_qcpool()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-median_bc"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-median_bc}{}}}
\subsection{Method \code{median_bc()}}{
Median batch correction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$median_bc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame with median batch corrected data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-loess_bc"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-loess_bc}{}}}
\subsection{Method \code{loess_bc()}}{
Perform QC-RLSC for batch correction of chromatographic signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$loess_bc(tab, colv, or, span = 0.75, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tab}}{table N*K (row * column) with N samples and K variables.}

\item{\code{colv}}{vector N of numbers: 1 for QC samples and 2 for other samples.}

\item{\code{or}}{vector of measuring order (see details).}

\item{\code{span}}{the parameter alpha which controls the degree of smoothing.}

\item{\code{verbose}}{print which variable has been corrected to monitor the process (default = FALSE).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make sure that everything is sorted in measurement order!!!
}

\subsection{Returns}{
corrected table N*K
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-combat_bc"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-combat_bc}{}}}
\subsection{Method \code{combat_bc()}}{
Perform ComBat batch correction from the SVA package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$combat_bc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame with combat batch corrected data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-blank_filtering"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-blank_filtering}{}}}
\subsection{Method \code{blank_filtering()}}{
Perform sample blank / ratio filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$blank_filtering()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-plot_qc_rsd"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-plot_qc_rsd}{}}}
\subsection{Method \code{plot_qc_rsd()}}{
Show the histogram of RSD values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$plot_qc_rsd()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsd_limit}}{numeric(1), set a limit between 0 and 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object, RSD histogram.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-plot_qc_trend"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-plot_qc_trend}{}}}
\subsection{Method \code{plot_qc_trend()}}{
Show the trend plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$plot_qc_trend()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ggplot2 object, RSD histogram.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
