% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{di_fun5}
\alias{di_fun5}
\title{test function}
\usage{
di_fun5(self, private)
}
\arguments{
\item{self}{class object}

\item{private}{class object}
}
\value{
private
}
\description{
Can I change a private property from here?
}
