% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{extract_feature_data}
\alias{extract_feature_data}
\title{Extract feature data}
\usage{
extract_feature_data(self = NULL)
}
\arguments{
\item{self}{class object}
}
\value{
self (invisible)
}
\description{
Extract feature data.
}
