% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{extract_indices}
\alias{extract_indices}
\title{Extract indices from the data}
\usage{
extract_indices(self = self)
}
\arguments{
\item{self}{class object}
}
\value{
self (invisible)
}
\description{
Extract indices from the data.
}
