% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract tables from the data}
\usage{
extract_tables(self = self)
}
\arguments{
\item{self}{class object}
}
\value{
self (invisible)
}
\description{
Extract tables from the data.
}
