% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{import_read_metadata}
\alias{import_read_metadata}
\title{Import the meta data}
\usage{
import_read_metadata(self = NULL)
}
\arguments{
\item{self}{self}
}
\value{
self
}
\description{
Import the meta data.
}
