% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{make_table_long}
\alias{make_table_long}
\title{Make data.frame long}
\usage{
make_table_long(self = NULL)
}
\arguments{
\item{self}{data.frame in wide format.}
}
\value{
self (invisible).
}
\description{
Make data.frame long.
}
