% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{make_table_wide}
\alias{make_table_wide}
\title{Make data.frame wide}
\usage{
make_table_wide(data_long = NULL)
}
\arguments{
\item{data_long}{data.frame in long format.}
}
\value{
data.frame in wide format.
}
\description{
Make data.frame wide.
}
