% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{norm_pqn}
\alias{norm_pqn}
\title{PQN normalisation}
\usage{
norm_pqn(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible)
}
\description{
PQN normalisation.
}
