% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{norm_total_area}
\alias{norm_total_area}
\title{Total area normalisation}
\usage{
norm_total_area(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible)
}
\description{
Total area normalisation.
}
