% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_apply_rsd}
\alias{qc_apply_rsd}
\title{Apply the RSD filtering}
\usage{
qc_apply_rsd(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible).
}
\description{
Apply the RSD filtering.
}
