% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_calc_cor}
\alias{qc_calc_cor}
\title{Calculate the correlation for all samples and qcpools}
\usage{
qc_calc_cor(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible).
}
\description{
Calculate the correlation for all samples and qcpools.
}
