% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_calc_rsd}
\alias{qc_calc_rsd}
\title{Calculate the RSD values}
\usage{
qc_calc_rsd(self = NULL, private = NULL)
}
\arguments{
\item{self}{object of class DataImport.}

\item{private}{private part of object.}
}
\value{
self (invisible).
}
\description{
Calculate the RSD values of all features.
}
