% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_calc_trend}
\alias{qc_calc_trend}
\title{Calculate the trend of all features}
\usage{
qc_calc_trend(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible).
}
\description{
Calculate the trend of all features.
}
