% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_plot_class_rsd}
\alias{qc_plot_class_rsd}
\title{A violin plot showing the RSD of each lipid species per lipid class}
\usage{
qc_plot_class_rsd(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
ggplot2 object containing a violin plot.
}
\description{
A violin plot showing the RSD of each lipid species per lipid class.
}
