% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_plot_cor}
\alias{qc_plot_cor}
\title{Show correlation heatmap samples and qcpools}
\usage{
qc_plot_cor(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
ggplot2, correlation heatmap.
}
\description{
Show correlation heatmap samples and qcpools.
}
