% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_plot_rsd}
\alias{qc_plot_rsd}
\title{Show RSD histogram}
\usage{
qc_plot_rsd(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
ggplot2 object, RSD histogram.
}
\description{
Show the histogram of RSD values.
}
