% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_plot_trend}
\alias{qc_plot_trend}
\title{Show the trend plot}
\usage{
qc_plot_trend(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
ggplot2 object, trend plot.
}
\description{
Show the trend plot.
}
