% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{read_msdial}
\alias{read_msdial}
\title{Read a MSDIAL report file}
\usage{
read_msdial(file = NULL)
}
\arguments{
\item{file}{character(1), full path to result file.}
}
\value{
data.frame with MSDIAL results.
}
\description{
Read a MSDIAL report file.
}
