% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UntargetedLipidomics.R
\name{UntargetedLipidomics}
\alias{UntargetedLipidomics}
\title{Defining the Untargeted lipidomics class}
\description{
Defining the Untargeted lipidomics class
}
\author{
Rico Derks

Yassene Mohammed
}
