% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_add_log}
\alias{utils_add_log}
\title{Add a log entry}
\usage{
utils_add_log(self = NULL, message = NULL)
}
\arguments{
\item{self}{self}

\item{message}{character(1), log message.}
}
\value{
self
}
\description{
Add log entry into the history.
}
