% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_analysis_features}
\alias{utils_analysis_features}
\title{Determine which features to keep}
\usage{
utils_analysis_features(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invsible).
}
\description{
Determine which features to keep depending on the filters used.
}
