% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_analysis_table}
\alias{utils_analysis_table}
\title{Extract the sample data for further analysis}
\usage{
utils_analysis_table(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible).
}
\description{
Exctract the sample data table for further analysis. All set filtering 
steps have been applied on this table.
}
