% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_file_show}
\alias{utils_file_show}
\title{Show the files from DataImport class}
\usage{
utils_file_show(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\description{
Show the files from DataImport class.
}
