% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_reset_tables}
\alias{utils_reset_tables}
\title{Reset the analysis table and feature table}
\usage{
utils_reset_tables(self = NULL)
}
\arguments{
\item{self}{object of class DataImport.}
}
\value{
self (invisible).
}
\description{
Reset the analysis table and feature table.
}
