% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_sample_show}
\alias{utils_sample_show}
\title{Show the samples from DataImport class}
\usage{
utils_sample_show(self = NULL)
}
\arguments{
\item{self}{self}
}
\description{
Show the samples from DataImport class.
}
