% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImport.R
\name{DataImport}
\alias{DataImport}
\title{R6 Data import class}
\description{
Data import class. This will be the parent class for several other classes.
}
\references{
Dunn et al. Nature Protocols 6, 1060-1083 (2011)
}
\author{
Rico Derks

Yassene Mohammed

E. Nevedomskaya
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character(1), containing a name.}

\item{\code{filenames}}{character(), containing the full file names.}

\item{\code{meta_filename}}{character(1), containing the full meta file name.}

\item{\code{tables}}{list() containing all data tables.}

\item{\code{indices}}{list() containing all kind of indices.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataImport-new}{\code{DataImport$new()}}
\item \href{#method-DataImport-print}{\code{DataImport$print()}}
\item \href{#method-DataImport-read_meta_data}{\code{DataImport$read_meta_data()}}
\item \href{#method-DataImport-make_table_long}{\code{DataImport$make_table_long()}}
\item \href{#method-DataImport-make_table_wide}{\code{DataImport$make_table_wide()}}
\item \href{#method-DataImport-median_bc}{\code{DataImport$median_bc()}}
\item \href{#method-DataImport-loess_bc}{\code{DataImport$loess_bc()}}
\item \href{#method-DataImport-combat_bc}{\code{DataImport$combat_bc()}}
\item \href{#method-DataImport-plot_qc_rsd}{\code{DataImport$plot_qc_rsd()}}
\item \href{#method-DataImport-plot_qc_trend}{\code{DataImport$plot_qc_trend()}}
\item \href{#method-DataImport-clone}{\code{DataImport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-new}{}}}
\subsection{Method \code{new()}}{
Initialization function for DataImport class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$new(name = NA, filenames = NA, meta_filename = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character(1), name.}

\item{\code{filenames}}{character(), containing the full file names.}

\item{\code{meta_filename}}{character(1), containing the full meta file name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-print}{}}}
\subsection{Method \code{print()}}{
Print a nice summary of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used at the moment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-read_meta_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-read_meta_data}{}}}
\subsection{Method \code{read_meta_data()}}{
Read all kind of meta data files (.csv, .txt, .xlsx).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$read_meta_data(file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{character(1), containing the full path to the meta data file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame containing the meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-make_table_long"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-make_table_long}{}}}
\subsection{Method \code{make_table_long()}}{
Function to pivot a data.frame into long (tidy) format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$make_table_long()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame in long format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-make_table_wide"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-make_table_wide}{}}}
\subsection{Method \code{make_table_wide()}}{
Function to pivot a data.frame into wide format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$make_table_wide()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame in wide format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-median_bc"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-median_bc}{}}}
\subsection{Method \code{median_bc()}}{
Median batch correction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$median_bc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame with median batch corrected data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-loess_bc"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-loess_bc}{}}}
\subsection{Method \code{loess_bc()}}{
Perform QC-RLSC for batch correction of chromatographic signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$loess_bc(tab, colv, or, span = 0.75, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tab}}{table N*K (row * column) with N samples and K variables.}

\item{\code{colv}}{vector N of numbers: 1 for QC samples and 2 for other samples.}

\item{\code{or}}{vector of measuring order (see details).}

\item{\code{span}}{the parameter alpha which controls the degree of smoothing.}

\item{\code{verbose}}{print which variable has been corrected to monitor the process (default = FALSE).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make sure that everything is sorted in measurement order!!!
}

\subsection{Returns}{
corrected table N*K
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-combat_bc"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-combat_bc}{}}}
\subsection{Method \code{combat_bc()}}{
Perform ComBat batch correction from the SVA package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$combat_bc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame with combat batch corrected data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-plot_qc_rsd"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-plot_qc_rsd}{}}}
\subsection{Method \code{plot_qc_rsd()}}{
A histogram showing the RSD distribution of all features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$plot_qc_rsd()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ggplot2 object, containing a histogram.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-plot_qc_trend"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-plot_qc_trend}{}}}
\subsection{Method \code{plot_qc_trend()}}{
Trend plot showing how each feature behaves in the pooled samples compared 
to the first pooled sample in the batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$plot_qc_trend()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ggplot2 object, containing a trend plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataImport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataImport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataImport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
