% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{make_table_wide}
\alias{make_table_wide}
\title{Make data.frame wide}
\usage{
make_table_wide(self = NULL)
}
\arguments{
\item{self}{data.frame in long format.}
}
\value{
self (invisible).
}
\description{
Make data.frame wide.
}
