% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UntargetedLipidomics.R
\name{UntargetedLipidomics}
\alias{UntargetedLipidomics}
\title{R6 untargeted lipidomics class}
\description{
Untargeted lipidomics class. This is a child class from DataImport class.
}
\author{
Rico Derks

Yassene Mohammed
}
\section{Super class}{
\code{\link[TomatoR6:DataImport]{TomatoR6::DataImport}} -> \code{UntargetedLipidomics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UntargetedLipidomics-new}{\code{UntargetedLipidomics$new()}}
\item \href{#method-UntargetedLipidomics-plot_qc_rsd_class}{\code{UntargetedLipidomics$plot_qc_rsd_class()}}
\item \href{#method-UntargetedLipidomics-clone}{\code{UntargetedLipidomics$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="combat_bc"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-combat_bc'><code>TomatoR6::DataImport$combat_bc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="loess_bc"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-loess_bc'><code>TomatoR6::DataImport$loess_bc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="make_table_long"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-make_table_long'><code>TomatoR6::DataImport$make_table_long()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="make_table_wide"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-make_table_wide'><code>TomatoR6::DataImport$make_table_wide()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="median_bc"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-median_bc'><code>TomatoR6::DataImport$median_bc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="plot_qc_rsd"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-plot_qc_rsd'><code>TomatoR6::DataImport$plot_qc_rsd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="plot_qc_trend"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-plot_qc_trend'><code>TomatoR6::DataImport$plot_qc_trend()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="print"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-print'><code>TomatoR6::DataImport$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="read_meta_data"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-read_meta_data'><code>TomatoR6::DataImport$read_meta_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UntargetedLipidomics-new"></a>}}
\if{latex}{\out{\hypertarget{method-UntargetedLipidomics-new}{}}}
\subsection{Method \code{new()}}{
Initialization function for DataImport class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UntargetedLipidomics$new(name = NA, filenames = NA, meta_filename = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character(1), name.}

\item{\code{filenames}}{character(), containing the full file names.}

\item{\code{meta_filename}}{character(1), containing the full meta file name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UntargetedLipidomics-plot_qc_rsd_class"></a>}}
\if{latex}{\out{\hypertarget{method-UntargetedLipidomics-plot_qc_rsd_class}{}}}
\subsection{Method \code{plot_qc_rsd_class()}}{
A violin plot showing the RSD of each lipid species per lipid class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UntargetedLipidomics$plot_qc_rsd_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ggplot2 object containing a violin plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UntargetedLipidomics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UntargetedLipidomics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UntargetedLipidomics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
