% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UntargetedMetabolomics.R
\name{UntargetedMetabolomics}
\alias{UntargetedMetabolomics}
\title{R6 untargeted metabolomics class}
\description{
Untargeted metabolomics class. This is a child class from DataImport class.
}
\author{
Rico Derks

Yassene Mohammed
}
\section{Super class}{
\code{\link[TomatoR6:DataImport]{TomatoR6::DataImport}} -> \code{UntargetedMetabolomics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UntargetedMetabolomics-new}{\code{UntargetedMetabolomics$new()}}
\item \href{#method-UntargetedMetabolomics-clone}{\code{UntargetedMetabolomics$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="fun1"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-fun1'><code>TomatoR6::DataImport$fun1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="fun2"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-fun2'><code>TomatoR6::DataImport$fun2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="fun3"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-fun3'><code>TomatoR6::DataImport$fun3()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TomatoR6" data-topic="DataImport" data-id="fun4"><a href='../../TomatoR6/html/DataImport.html#method-DataImport-fun4'><code>TomatoR6::DataImport$fun4()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UntargetedMetabolomics-new"></a>}}
\if{latex}{\out{\hypertarget{method-UntargetedMetabolomics-new}{}}}
\subsection{Method \code{new()}}{
Initialization function for DataImport class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UntargetedMetabolomics$new(name = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character(1), name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UntargetedMetabolomics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UntargetedMetabolomics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UntargetedMetabolomics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
