% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UntargetedMetabolomics.R
\name{UntargetedMetabolomics}
\alias{UntargetedMetabolomics}
\title{Defining the Untargeted metabolomics class}
\description{
Defining the Untargeted metabolomics class
}
\author{
Rico Derks

Yassene Mohammed
}
